move import_bgen to testutils
