+++ hail/src/test/scala/is/hail/expr/ir/IRSuite.scala
@@ -749,7 +749,7 @@ class IRSuite extends SparkSuite {
     val mt = MatrixTable.range(hc, 20, 2, Some(3)).ast.asInstanceOf[MatrixRead]
     val vcf = is.hail.TestUtils.importVCF("src/test/resources/sample.vcf")
       .ast.asInstanceOf[MatrixRead]
-    val bgen = hc.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
+    val bgen = is.hail.TestUtils.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
       .ast.asInstanceOf[MatrixRead]
 
     val irs = Array(
@@ -886,7 +886,7 @@ class IRSuite extends SparkSuite {
         .ast.asInstanceOf[MatrixRead]
       val vcf = is.hail.TestUtils.importVCF("src/test/resources/sample.vcf")
         .ast.asInstanceOf[MatrixRead]
-      val bgen = hc.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
+      val bgen = is.hail.TestUtils.importBgens(FastIndexedSeq("src/test/resources/example.8bits.bgen"))
         .ast.asInstanceOf[MatrixRead]
       val range1 = MatrixTable.range(hc, 20, 2, Some(3))
         .ast.asInstanceOf[MatrixRead]
